+++ clox/object.c
@@ -31,7 +31,7 @@ static ObjString* allocateString(char* chars, int length, uint32_t hash) {
 static uint32_t hashString(const char* key, int length) {
     uint32_t hash = 2166136261u;
     for (int i = 0; i < length; i++) {
-        hash ^= (uint8_t)key[i];
+        hash ^= (uint16_t)key[i];
         hash *= 16777619;
     }
